Move apiV3 config into run block
